+++ src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/EndpointContracts/RazorLanguageQueryParams.cs
@@ -1,12 +1,15 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the MIT license. See License.txt in the project root for license information.
 
-using MediatR;
+using System;
 using Microsoft.VisualStudio.LanguageServer.Protocol;
 
-namespace Microsoft.AspNetCore.Razor.LanguageServer.EndpointContracts
+namespace Microsoft.AspNetCore.Razor.LanguageServer
 {
-    internal class DidChangeConfigurationParamsBridge : DidChangeConfigurationParams, IRequest
+    internal class RazorLanguageQueryParams
     {
+        public required Uri Uri { get; set; }
+
+        public required Position Position { get; set; }
     }
 }
