+++ src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/RazorLanguageQueryEndpoint.cs
@@ -1,0 +1,93 @@
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the MIT license. See License.txt in the project root for license information.
+
+using System.Threading;
+using System.Threading.Tasks;
+using Microsoft.AspNetCore.Razor.LanguageServer.Common.Extensions;
+using Microsoft.AspNetCore.Razor.LanguageServer.EndpointContracts;
+using Microsoft.AspNetCore.Razor.LanguageServer.Protocol;
+using Microsoft.CodeAnalysis.Text;
+using Microsoft.VisualStudio.LanguageServer.Protocol;
+
+namespace Microsoft.AspNetCore.Razor.LanguageServer
+{
+    internal class RazorLanguageQueryEndpoint : IRazorLanguageQueryHandler
+    {
+        public bool MutatesSolutionState => false;
+
+        private readonly RazorDocumentMappingService _documentMappingService;
+
+        public RazorLanguageQueryEndpoint(RazorDocumentMappingService documentMappingService)
+        {
+            _documentMappingService = documentMappingService;
+        }
+
+        public TextDocumentIdentifier GetTextDocumentIdentifier(RazorLanguageQueryParams request)
+        {
+            return new TextDocumentIdentifier
+            {
+                Uri = request.Uri
+            };
+        }
+
+        public async Task<RazorLanguageQueryResponse> HandleRequestAsync(RazorLanguageQueryParams request, RazorRequestContext requestContext, CancellationToken cancellationToken)
+        {
+            requestContext.RequireDocumentContext();
+            var documentContext = requestContext.DocumentContext;
+
+            var documentSnapshot = documentContext.Snapshot;
+            var documentVersion = documentContext.Version;
+
+            var codeDocument = await documentSnapshot.GetGeneratedOutputAsync();
+            var sourceText = await documentSnapshot.GetTextAsync();
+            var linePosition = new LinePosition(request.Position.Line, request.Position.Character);
+            var hostDocumentIndex = sourceText.Lines.GetPosition(linePosition);
+            var responsePosition = request.Position;
+
+            if (codeDocument.IsUnsupported())
+            {
+                // All language queries on unsupported documents return Html. This is equivalent to what pre-VSCode Razor was capable of.
+                return new RazorLanguageQueryResponse()
+                {
+                    Kind = RazorLanguageKind.Html,
+                    Position = responsePosition,
+                    PositionIndex = hostDocumentIndex,
+                    HostDocumentVersion = documentVersion,
+                };
+            }
+
+            var responsePositionIndex = hostDocumentIndex;
+
+            var languageKind = _documentMappingService.GetLanguageKind(codeDocument, hostDocumentIndex, rightAssociative: false);
+            if (languageKind == RazorLanguageKind.CSharp)
+            {
+                if (_documentMappingService.TryMapToProjectedDocumentPosition(codeDocument, hostDocumentIndex, out var projectedPosition, out var projectedIndex))
+                {
+                    // For C# locations, we attempt to return the corresponding position
+                    // within the projected document
+                    responsePosition = projectedPosition;
+                    responsePositionIndex = projectedIndex;
+                }
+                else
+                {
+                    // It no longer makes sense to think of this location as C#, since it doesn't
+                    // correspond to any position in the projected document. This should not happen
+                    // since there should be source mappings for all the C# spans.
+                    languageKind = RazorLanguageKind.Razor;
+                    responsePositionIndex = hostDocumentIndex;
+                }
+            }
+
+            requestContext.LspLogger.LogInformation("Language query request for ({requestPositionLine}, {requestPositionCharacter}) = {languageKind} at ({responsePositionLine}, {responsePositionCharacter})",
+                request.Position.Line, request.Position.Character, languageKind, responsePosition.Line, responsePosition.Character);
+
+            return new RazorLanguageQueryResponse()
+            {
+                Kind = languageKind,
+                Position = responsePosition,
+                PositionIndex = responsePositionIndex,
+                HostDocumentVersion = documentVersion
+            };
+        }
+    }
+}
